figma-transformer: extract CSS positioning resolver
